Fix 2d_em_uniform pic demo test
Update the 2d_em_uniform demo_test to properly import the
Plotter object and call correct methods to plot velocity.
Adding gridlines to animator class and change params to
allow non-integer nppc.

+++ py_platypus/src/py_platypus/utils/params.py
@@ -93,7 +93,7 @@ class Parameters:
         self.graph_dir = os.path.join(self.out_dir, "graph")
         self.set("dx", list(map(lambda a, b: a/b, self["length"], self["cells"])), refresh=False)
         self.set("steps", int(self["runtime"]/self["timestep"]), refresh=False)
-        self.set("n_particles", self["nppc"] * np.prod(self["cells"]), refresh=False)
+        self.set("n_particles", int(self["nppc"] * np.prod(self["cells"])), refresh=False)
 
         return
